refactor: create dynamoDB service
