+++ product-service/handlers/getProductsList/index.js
@@ -1,27 +1,11 @@
-import AWS from "aws-sdk";
-
+import { getProducts } from "../../services/dynamoDB";
 import { logger } from "../../services/Logger";
-
-const dynamo = new AWS.DynamoDB.DocumentClient();
 
 export const getProductsList = async () => {
   try {
     logger.info("START getProductsList");
 
-    const [productsScan, stocksScan] = await Promise.all([
-      dynamo.scan({ TableName: process.env.PRODUCTS_TABLE_NAME }).promise(),
-      dynamo.scan({ TableName: process.env.STOCKS_TABLE_NAME }).promise(),
-    ]);
-
-    const productIdToCountMap = stocksScan.Items.reduce((acc, next) => {
-      acc[next.product_id] = next.count;
-      return acc;
-    }, {});
-
-    const productsWithCount = productsScan.Items.map((product) => ({
-      ...product,
-      count: productIdToCountMap[product.id] ?? 0,
-    }));
+    const products = await getProducts();
 
     logger.info("FINISH SUCCESS getProductsList");
     return {
@@ -30,7 +14,7 @@ export const getProductsList = async () => {
         "Access-Control-Allow-Origin": "*",
         "Access-Control-Allow-Credentials": true,
       },
-      body: JSON.stringify(productsWithCount ?? []),
+      body: JSON.stringify(products),
     };
   } catch (error) {
     logger.error(error);
